requirements.txt
piexif==1.1.3
Pillow==8.3.1
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
